Add a Riak configuration function (#3)
* Add function to extract logger config

The aim is to support the config from previous Riak, the classic riak config, the addition of a json log, and domain filtering - all through one config fetch function.

It should be flexible for other extensions - or other functions/modules could be added.

* Use better config names

* Align additional_handler and default_filters with riak.conf output

It is assumes there will be cuttlefish validators, checking that only a pipe-delimited list with supported strings (that convert to supported atoms).

* Config will now use list datatype in cuttlefish

... so need to split a delimited string

* Have `none` as a default - and strip to empty list